Canceled PythonBot online
It was unfinished, and thinking about it the use case was not that interesting.
Removed all that was related to running the game view in a browser. The engine
can still send the state of the game on a socket, this has not been removed.

This can be enabled by providing the argument `-s` to the engine (would be done
in `pythonbot.py`), and opening a socket which listens for the status of the
game sent by the engine.

The directory WebServer/ has been removed, and the content of the directory
Engine/ moved back to the root of the repository.